requirements.txt
flask
flask-cors
python-dotenv
langgraph
langchain-core
langchain-google-genai
